+++ sys/arch/xen/include/intr.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: intr.h,v 1.13 2006/12/21 15:55:25 yamt Exp $	*/
+/*	$NetBSD: intr.h,v 1.14 2006/12/26 15:22:44 ad Exp $	*/
 /*	NetBSD intr.h,v 1.15 2004/10/31 10:39:34 yamt Exp	*/
 
 /*-
@@ -180,7 +180,7 @@ spllower(int nlevel)
 #define	spl0()		spllower(IPL_NONE)
 #define	splx(x)		spllower(x)
 
-typedef int ipl_t;
+typedef uint8_t ipl_t;
 typedef struct {
 	ipl_t _ipl;
 } ipl_cookie_t;
